pyproject.toml
[project]
name = "performance-tracker"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "httpx>=0.28.1",
    "pandas>=2.3.3",
    "pandas-stubs~=2.3.3",
    "python-dotenv>=1.2.1",
    "streamlit>=1.52.2",
    "yfinance>=1.0",
    "cachetools>=6.2.4",
    "python-dotenv>=1.2.1",
    "google-genai>=1.56.0",
    "pydantic>=2.12.5",
]

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[dependency-groups]
dev = [
    "coverage>=7.13.2",
    "coverage-badge>=1.1.2",
    "pre-commit>=4.5.1",
    "ruff>=0.14.10",
]

[tool.pyright]
include = ["src"]
typeCheckingMode = "basic"
pythonVersion = "3.13"

[tool.ruff]
line-length = 88
target-version = "py313"

[tool.ruff.lint]
ignore = [
    "E501",
    "ANN101",
    "ANN102",
    "ANN003",
    "PLR0913"
]
select = [
    "E",
    "W",
    "F",
    "I",
    "B",
    "SIM",
    "ANN",
    "PLR",
    "PLE",
    "PLW",
    "PLC",
    "UP",
    "N",
    "PTH",
    "PD"
]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
docstring-code-format = true
docstring-code-line-length = 72

[tool.coverage.run]
dynamic_context = "test_function"

[tool.coverage.html]
show_contexts = true
